final state, whitespace

+++ ch02/math_float.c
@@ -1,10 +1,12 @@
 #include <stdio.h>
+
 void main (void)
 {
   float a;
   float b = 3.641;
-  int c;
+  float c;
+
   a = 2.897;
   c = a + b;
-  printf ("The sum of adding %f and %f is %d\n", a, b, c);
+  printf ("The sum of adding %f and %f is %f\n", a, b, c);
 }
